feat(kyc): prevent showing flow if user has been rejected
