+++ packages/blockchain-wallet-v4-frontend/src/modals/Brokerage/Banks/Withdraw/selectors.ts
@@ -1,0 +1,13 @@
+import { lift } from 'ramda'
+
+import { ExtractSuccess } from '@core/types'
+import { createDeepEqualSelector } from '@core/utils'
+import { selectors } from 'data'
+
+const getData = createDeepEqualSelector([selectors.modules.profile.getUserData], (userDataR) => {
+  return lift((userData: ExtractSuccess<typeof userDataR>) => ({
+    userData
+  }))(userDataR)
+})
+
+export default getData
