+++ packages/blockchain-wallet-v4-frontend/src/modals/BuySell/index.tsx
@@ -22,6 +22,7 @@ import { BankStatusType, FastLinkType, ModalName } from 'data/types'
 import ModalEnhancer from 'providers/ModalEnhancer'
 
 import { Loading as StdLoading, LoadingTextEnum } from '../components'
+import Rejected from '../components/Rejected'
 import { ModalPropsType } from '../types'
 import AddCardCheckoutDotCom from './AddCardCheckoutDotCom'
 import AddCardEverypay from './AddCardEverypay'
@@ -42,7 +43,6 @@ import getData from './selectors'
 import SellOrderSummary from './SellOrderSummary'
 import Loading from './template.loading'
 import Pending from './template.pending'
-import Rejected from './template.rejected'
 import ThreeDSHandlerCheckoutDotCom from './ThreeDSHandlerCheckoutDotCom'
 import ThreeDSHandlerEverypay from './ThreeDSHandlerEverypay'
 import ThreeDSHandlerStripe from './ThreeDSHandlerStripe'
